app/services.py
import boto3
import os
from dotenv import load_dotenv

# Load environment variables from the .env file
load_dotenv()

# Initialize Cognito client
cognito_client = boto3.client(
    'cognito-idp',
    region_name=os.getenv('AWS_REGION'),
    aws_access_key_id=os.getenv('AWS_ACCESS_KEY_ID'),
    aws_secret_access_key=os.getenv('AWS_SECRET_ACCESS_KEY')
)

# Cognito app client ID and user pool ID from environment variables
COGNITO_APP_CLIENT_ID = os.getenv('COGNITO_APP_CLIENT_ID')
COGNITO_USER_POOL_ID = os.getenv('COGNITO_USER_POOL_ID')