requirements.txt
aniso8601==10.0.1
blinker==1.9.0
boto3==1.38.10
botocore==1.38.10
click==8.1.8
Flask==3.1.0
Flask-Cors==3.0.10
Flask-RESTful==0.3.10
Flask-S3==0.3.3
Flask-WTF==1.2.2
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
MarkupSafe==3.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
s3transfer==0.12.0
six==1.17.0
urllib3==2.4.0
Werkzeug==3.1.3
WTForms==3.2.1